requirements.txt
fastapi==0.118.0
numba==0.61.2
numpy==1.26.4
pandas==2.3.3
pyserial==3.5
Requests==2.32.5
scipy==1.16.2
uvicorn==0.37.0
